Add files via upload
Bash script that curls a list of URLs from a text file and writes out either only status code or status code and headers depending on script of choise. Name input file domains.txt and run either one of the scripts. Wrote it due to the problems ive had with httpx lately.